requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.4
django-cors-headers==4.3.1
django-cron==0.6.0
django-dotenv==1.4.2
django-embed-video==1.4.9
django-environ==0.11.2
django-picklefield==3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
httplib2==0.22.0
idna==3.6
jsonfield==3.1.0
numpy==1.26.4
pandas==2.2.1
pillow==10.3.0
PyJWT==2.8.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
razorpay==1.4.2
requests==2.31.0
schedule==1.2.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
